Refactor arc path creation w/ existing StatesLayer
