+++ src/app/src/components/AnimatedArcsAndMap/useCreateArcPath.tsx
@@ -1,60 +1,31 @@
-import { useEffect, useRef, useCallback } from 'react';
+import { useCallback } from 'react';
 import { useMap } from 'react-leaflet';
-
 import L from 'leaflet';
-import '@elfalem/leaflet-curve';
-import { StateGeometry, StateProperties } from '../states.geojson';
+import { useGetSpendingOverTimeQuery } from '../../api';
+import { MonthlySpendingOverTime } from '../../types/api';
+import { PROGRESS_FINAL_STEP } from '../../util';
+import { DC_CENTER, MONTHLY_TIME_DURATION } from '../../constants';
 
-import StatesLayer from '../StatesLayer';
-import {
-    MonthlySpendingOverTime,
-    MonthlySpendingOverTimeByState,
-} from '../../types/api';
-import { MONTHLY_TIME_DURATION, DC_CENTER } from '../../constants';
-
-function getBezierOffsetLatLng(end: L.LatLngTuple): L.LatLngTuple {
-    // Modified Bezier curve function from https://gist.github.com/ryancatalani/6091e50bf756088bf9bf5de2017b32e6
-    const start: L.LatLngTuple = DC_CENTER;
-    var offsetX = end[1] - start[1],
-        offsetY = end[0] - start[0];
-
-    var r = Math.sqrt(Math.pow(offsetX, 2) + Math.pow(offsetY, 2)),
-        theta = Math.atan2(offsetY, offsetX);
-
-    var thetaOffset = 3.14 / 10;
-
-    var r2 = r / 2 / Math.cos(thetaOffset),
-        theta2 = theta + thetaOffset;
-
-    var midpointX = r2 * Math.cos(theta2) + start[1],
-        midpointY = r2 * Math.sin(theta2) + start[0];
-
-    return [midpointY, midpointX];
-}
-
-export default function AnimatedArcsOverStates({
-    animationEnabled,
-    spending,
-    totalTimeSteps,
-}: {
-    animationEnabled: boolean;
-    spending: MonthlySpendingOverTimeByState;
-    totalTimeSteps: number;
-}) {
+export default function useCreateArcPath(
+    arcPathsReference: React.MutableRefObject<
+        {
+            shape_code: string;
+            curve: L.Curve;
+        }[]
+    >
+) {
+    const { data: spending } = useGetSpendingOverTimeQuery();
     const map = useMap();
-    const arcPathsReference = useRef<{ shape_code: string; curve: L.Curve }[]>(
-        []
-    );
-
     !map.getPane('arcPathsPane') && map.createPane('arcPathsPane');
 
-    const createArcPath = useCallback(
+    return useCallback(
         (event: any) => {
             const state = event.sourceTarget.feature.properties.STUSPS;
             if (
                 arcPathsReference.current.find(
                     path => path.shape_code === state
-                )
+                ) ||
+                !spending
             ) {
                 return;
             }
@@ -72,7 +43,7 @@ export default function AnimatedArcsOverStates({
                 spendingForState: spending.find(
                     data => data.shape_code === state
                 )!,
-                totalTimeSteps: totalTimeSteps,
+                totalTimeSteps: PROGRESS_FINAL_STEP,
             });
 
             arcPathsReference.current.push({
@@ -92,33 +63,7 @@ export default function AnimatedArcsOverStates({
                 ),
             });
         },
-        [arcPathsReference, spending, totalTimeSteps]
-    );
-
-    useEffect(() => {
-        if (!animationEnabled && arcPathsReference.current.length > 0) {
-            arcPathsReference.current.forEach(path => {
-                path.curve.removeFrom(map);
-            });
-            arcPathsReference.current = [];
-        }
-        animationEnabled &&
-            arcPathsReference.current.forEach(path => {
-                path.curve.addTo(map);
-            });
-    }, [map, animationEnabled]);
-
-    return (
-        <>
-            <StatesLayer
-                onEachFeature={(
-                    feature,
-                    layer: L.GeoJSON<StateProperties, StateGeometry>
-                ) => {
-                    layer.on('add', createArcPath);
-                }}
-            />
-        </>
+        [arcPathsReference, spending]
     );
 }
 
@@ -158,4 +103,24 @@ function getPathAnimationValues({
         startMonth: startAtTime,
         playDuration: totalTime - startAtTime,
     };
+}
+
+function getBezierOffsetLatLng(end: L.LatLngTuple): L.LatLngTuple {
+    // Modified Bezier curve function from https://gist.github.com/ryancatalani/6091e50bf756088bf9bf5de2017b32e6
+    const start: L.LatLngTuple = DC_CENTER;
+    var offsetX = end[1] - start[1],
+        offsetY = end[0] - start[0];
+
+    var r = Math.sqrt(Math.pow(offsetX, 2) + Math.pow(offsetY, 2)),
+        theta = Math.atan2(offsetY, offsetX);
+
+    var thetaOffset = 3.14 / 10;
+
+    var r2 = r / 2 / Math.cos(thetaOffset),
+        theta2 = theta + thetaOffset;
+
+    var midpointX = r2 * Math.cos(theta2) + start[1],
+        midpointY = r2 * Math.sin(theta2) + start[0];
+
+    return [midpointY, midpointX];
 }
